clean, safe interface for instance (so that preproc is safe)
